requirements.txt
streamlit
speechrecognition
langchain
faiss-cpu
sentence-transformers
google-generativeai
pypdf
